+++ main.h
@@ -38,7 +38,7 @@ void negative_octal_div(unsigned n);
 void negative_hex_div(unsigned n, int flag);
 void binary_div(unsigned n);
 void octal_div(unsigned n);
-void hex_div(unsigned n, int flag);
+void hex_div(unsigned long n, int flag);
 int print_hex_upper(va_list list);
 int print_hex_lower(va_list list);
 char letter(unsigned n, int flag);
